abc043

+++ abc043.py
@@ -1,5 +1,5 @@
 n = int(input())
 
-answer = (n + 1) * n / 2
+answer = int((n + 1) * n / 2)
 
 print(answer)
